\name{panel.assoc}
\alias{panel.assoc}
\title{Panel Function for Association Plots}
\description{
  A concise (1-5 lines) description of what the function does.
}
\usage{
panel.assoc(x, i, j, type, legend = FALSE, axes = TRUE,
  margins = c(0, 0, 0, 0), abbreviate = FALSE, gp = NULL,
  labeling = NULL, shade = FALSE, ...)
}
\arguments{
  \item{x}{What is \code{x}?}
  \item{i}{What is \code{i}?}
  \item{j}{What is \code{j}?}
  \item{type}{What is \code{type}?}
  \item{legend}{What is \code{legend}?}
  \item{axes}{What is \code{axes}?}
  \item{margins}{What is \code{margins}?}
  \item{abbreviate}{What is \code{abbreviate}?}
  \item{gp}{What is \code{gp}?}
  \item{labeling}{What is \code{labeling}?}
  \item{shade}{What is \code{shade}?}
  \item{\dots}{What is \code{\dots}?}
}
\details{
  more details
}
\value{
  Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}
  ...
}

\seealso{\code{\link{strucplot}}}
\examples{

}
\keyword{hplot}
