\name{Baseball}
\alias{Baseball}
\non_function{}
\title{Baseball}
\description{
Baseball data.
}

\usage{data(Baseball)}
\format{A data frame with 322 observations and 25 variables.
  \describe{
    \item{name1}{factor indicating ,}
    \item{name2}{factor indicating ,}
    \item{atbat86}{,}
    \item{hits86}{,}
    \item{homer86}{,}
    \item{runs86}{,}
    \item{rbi86}{,}
    \item{walks86}{,}
    \item{years}{,}
    \item{atbat}{,}
    \item{hits}{,}
    \item{homeruns}{,}
    \item{runs}{,}
    \item{rbi}{,}
    \item{walks}{,}
    \item{league86}{,}
    \item{div86}{,}
    \item{team86}{,}
    \item{posit86}{,}
    \item{outs86}{,}
    \item{assist86}{,}
    \item{error86}{,}
    \item{sal87}{,}
    \item{league87}{,}
    \item{team87}{.}
  }
}

\references{
M. Friendly (2000),
\emph{Visualizing Categorical Data}. SAS Publishing.
}

\source{Michael Friendly (2000), Visualizing Categorical Data, p. ??.}

\examples{
data(Baseball)
}

\keyword{datasets}
